Unify smooth_knn C++ functions.
